Electrode list no longer in state. Fixed some more layout bugs.
