+++ src/views/AverageWaveforms/WaveformWidget/sharedDrawnComponents/electrodeGeometryStateManagement.tsx
@@ -1,7 +1,6 @@
-import { Electrode } from 'contexts/RecordingSelectionContext'
 import { DragAction, dragReducer, DragState } from 'FigurlCanvas/DragCanvas'
 import { pointSpanToRegion, Vec2 } from "FigurlCanvas/Geometry"
-import { LayoutMode, PixelSpaceElectrode } from './ElectrodeGeometry'
+import { Electrode, LayoutMode, PixelSpaceElectrode } from './ElectrodeGeometry'
 import { computeElectrodeLocations, getDraggedElectrodeIds, getElectrodeAtPoint, xMargin as xMarginDefault } from './electrodeGeometryLayout'
 
 export interface ElectrodeGeometryState {
@@ -111,7 +110,7 @@ export const electrodeGeometryReducer = (state: ElectrodeGeometryState, action: 
         const clickedId = getElectrodeAtPoint(state.convertedElectrodes, state.pixelRadius, action.point)
         // Suppose we clicked nothing. If no modifier was down & there's a selection to clear, clear it;
         // otherwise nothing actually happens so return the existing state.
-        if (clickedId === 0 || clickedId === undefined) {
+        if (clickedId === undefined) {
             return (!(action.shift || action.ctrl) && action.selectedElectrodeIds.length > 0) 
             ?   {
                     ...state,
